Split out `validateMediaItem`
